+++ app/Domains/Auth/Events/Role/RoleDeleted.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace App\Domains\Auth\Events\Role;
+
+use App\Domains\Auth\Models\Role;
+use Illuminate\Queue\SerializesModels;
+
+/**
+ * Class RoleDeleted.
+ */
+class RoleDeleted
+{
+    use SerializesModels;
+
+    /**
+     * @var
+     */
+    public $role;
+
+    /**
+     * @param $role
+     */
+    public function __construct(Role $role)
+    {
+        $this->role = $role;
+    }
+}