+++ app/Domains/Auth/Events/User/UserStatusChanged.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace App\Domains\Auth\Events\User;
+
+use App\Domains\Auth\Models\User;
+use Illuminate\Queue\SerializesModels;
+
+/**
+ * Class UserStatusChanged.
+ */
+class UserStatusChanged
+{
+    use SerializesModels;
+
+    /**
+     * @var
+     */
+    public $user;
+
+    /**
+     * @var
+     */
+    public $status;
+
+    /**
+     * UserStatusChanged constructor.
+     *
+     * @param  User  $user
+     * @param $status
+     */
+    public function __construct(User $user, $status)
+    {
+        $this->user = $user;
+        $this->status = $status;
+    }
+}