+++ app/Domains/Auth/Services/RoleService.php
@@ -2,6 +2,9 @@
 
 namespace App\Domains\Auth\Services;
 
+use App\Domains\Auth\Events\Role\RoleCreated;
+use App\Domains\Auth\Events\Role\RoleDeleted;
+use App\Domains\Auth\Events\Role\RoleUpdated;
 use App\Domains\Auth\Models\Role;
 use App\Exceptions\GeneralException;
 use App\Services\BaseService;
@@ -43,6 +46,8 @@ class RoleService extends BaseService
             throw new GeneralException(__('There was a problem creating the role.'));
         }
 
+        event(new RoleCreated($role));
+
         DB::commit();
 
         return $role;
@@ -69,6 +74,8 @@ class RoleService extends BaseService
             throw new GeneralException(__('There was a problem updating the role.'));
         }
 
+        event(new RoleUpdated($role));
+
         DB::commit();
 
         return $role;
@@ -86,6 +93,12 @@ class RoleService extends BaseService
             throw new GeneralException(__('You can not delete a role with associated users.'));
         }
 
-        return $this->deleteById($role->id);
+        if ($this->deleteById($role->id)) {
+            event(new RoleDeleted($role));
+
+            return true;
+        }
+
+        throw new GeneralException(__('There was a problem deleting the role.'));
     }
 }