+++ database/seeds/DatabaseSeeder.php
@@ -15,6 +15,7 @@ class DatabaseSeeder extends Seeder
         Model::unguard();
 
         $this->truncateMultiple([
+            'activity_log',
             'failed_jobs',
         ]);